+++ tests/test_document.py
@@ -1,11 +1,12 @@
 from __future__ import with_statement
+import bson
 import os
 import pickle
 import pymongo
-import bson
-import unittest
-import warnings
 import sys
+import unittest
+import uuid
+import warnings
 
 from nose.plugins.skip import SkipTest
 from datetime import datetime
@@ -1013,6 +1014,9 @@ class DocumentTest(unittest.TestCase):
 
         # Two posts with the same slug is not allowed
         post2 = BlogPost(title='test2', slug='test')
+        self.assertRaises(NotUniqueError, post2.save)
+
+        # Ensure backwards compatibilty for errors
         self.assertRaises(OperationError, post2.save)
 
     def test_unique_with(self):
@@ -1063,7 +1067,7 @@ class DocumentTest(unittest.TestCase):
 
         # Now there will be two docs with the same sub.slug
         post3 = BlogPost(title='test3', sub=SubDocument(year=2010, slug='test'))
-        self.assertRaises(OperationError, post3.save)
+        self.assertRaises(NotUniqueError, post3.save)
 
         BlogPost.drop_collection()
 
@@ -1090,11 +1094,11 @@ class DocumentTest(unittest.TestCase):
 
         # Now there will be two docs with the same sub.slug
         post3 = BlogPost(title='test3', sub=SubDocument(year=2010, slug='test'))
-        self.assertRaises(OperationError, post3.save)
+        self.assertRaises(NotUniqueError, post3.save)
 
         # Now there will be two docs with the same title and year
         post3 = BlogPost(title='test1', sub=SubDocument(year=2009, slug='test-1'))
-        self.assertRaises(OperationError, post3.save)
+        self.assertRaises(NotUniqueError, post3.save)
 
         BlogPost.drop_collection()
 
@@ -1117,7 +1121,7 @@ class DocumentTest(unittest.TestCase):
         try:
             cust_dupe.save()
             raise AssertionError, "We saved a dupe!"
-        except OperationError:
+        except NotUniqueError:
             pass
         Customer.drop_collection()
 
@@ -1642,6 +1646,52 @@ class DocumentTest(unittest.TestCase):
         person.reload()
         self.assertEqual(person.name, None)
         self.assertEqual(person.age, None)
+
+    def test_can_save_if_not_included(self):
+
+        class EmbeddedDoc(EmbeddedDocument):
+            pass
+
+        class Simple(Document):
+            pass
+
+        class Doc(Document):
+            string_field = StringField(default='1')
+            int_field = IntField(default=1)
+            float_field = FloatField(default=1.1)
+            boolean_field = BooleanField(default=True)
+            datetime_field = DateTimeField(default=datetime.now)
+            embedded_document_field = EmbeddedDocumentField(EmbeddedDoc, default=lambda: EmbeddedDoc())
+            list_field = ListField(default=lambda: [1, 2, 3])
+            dict_field = DictField(default=lambda: {"hello": "world"})
+            objectid_field = ObjectIdField(default=bson.ObjectId)
+            reference_field = ReferenceField(Simple, default=lambda: Simple().save())
+            map_field = MapField(IntField(), default=lambda: {"simple": 1})
+            decimal_field = DecimalField(default=1.0)
+            complex_datetime_field = ComplexDateTimeField(default=datetime.now)
+            url_field = URLField(default="http://mongoengine.org")
+            dynamic_field = DynamicField(default=1)
+            generic_reference_field = GenericReferenceField(default=lambda: Simple().save())
+            sorted_list_field = SortedListField(IntField(), default=lambda: [1, 2, 3])
+            email_field = EmailField(default="ross@example.com")
+            geo_point_field = GeoPointField(default=lambda: [1, 2])
+            sequence_field = SequenceField()
+            uuid_field = UUIDField(default=uuid.uuid4)
+            generic_embedded_document_field = GenericEmbeddedDocumentField(default=lambda: EmbeddedDoc())
+
+
+        Simple.drop_collection()
+        Doc.drop_collection()
+
+        Doc().save()
+
+        my_doc = Doc.objects.only("string_field").first()
+        my_doc.string_field = "string"
+        my_doc.save()
+
+        my_doc = Doc.objects.get(string_field="string")
+        self.assertEqual(my_doc.string_field, "string")
+        self.assertEqual(my_doc.int_field, 1)
 
     def test_document_update(self):